# swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login realizado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Credenciais inválidas"}},"summary":"Login com e-mail e senha","tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Usuário criado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"409":{"description":"E-mail já cadastrado"}},"summary":"Registrar novo usuário","tags":["Auth"]}},"/api/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"Perfil retornado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Não autorizado"}},"security":[{"JWT":[]}],"summary":"Obter perfil do usuário autenticado","tags":["Auth"]},"put":{"operationId":"AuthController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"Perfil atualizado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Não autorizado"}},"security":[{"JWT":[]}],"summary":"Atualizar perfil do usuário autenticado","tags":["Auth"]}},"/api/auth/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"Senha alterada com sucesso"},"400":{"description":"Senha atual incorreta"},"401":{"description":"Não autorizado"}},"security":[{"JWT":[]}],"summary":"Alterar senha do usuário autenticado","tags":["Auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"Token renovado"},"401":{"description":"Não autorizado"}},"security":[{"JWT":[]}],"summary":"Renovar token de acesso","tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"Dados do token"},"401":{"description":"Não autorizado"}},"security":[{"JWT":[]}],"summary":"Obter dados mínimos do token JWT","tags":["Auth"]}}},"info":{"title":"Yatsunami API","description":"API do sistema Yatsunami — gerenciamento de restaurante japonês","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Token JWT obtido via POST /api/auth/login"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com","description":"E-mail do usuário"},"password":{"type":"string","example":"123456","minLength":6,"description":"Senha do usuário"}},"required":["email","password"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"nome":{"type":"string","example":"João Silva"},"email":{"type":"string","example":"joao@example.com"},"role":{"type":"string","example":"user","enum":["user","admin"]},"tema":{"type":"string","example":"dark"},"idioma":{"type":"string","example":"pt-BR"}},"required":["id","nome","email","role","tema","idioma"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIs...","description":"JWT access token"},"user":{"$ref":"#/components/schemas/UserResponseDto"}},"required":["accessToken","user"]},"RegisterDto":{"type":"object","properties":{"nome":{"type":"string","example":"João Silva","description":"Nome completo"},"email":{"type":"string","example":"joao@example.com","description":"E-mail do usuário"},"password":{"type":"string","example":"123456","minLength":6,"description":"Senha"},"telefone":{"type":"string","example":"11999998888","description":"Telefone"},"cpfCnpj":{"type":"string","example":"123.456.789-00","description":"CPF ou CNPJ"}},"required":["nome","email","password"]},"UpdateProfileDto":{"type":"object","properties":{"nome":{"type":"string","example":"João Silva","description":"Nome completo"},"telefone":{"type":"string","example":"11999998888","description":"Telefone"},"tema":{"type":"string","example":"dark","enum":["light","dark","system"],"description":"Tema do app"},"idioma":{"type":"string","example":"pt-BR","enum":["pt-BR","ja-JP"],"description":"Idioma"},"endereco":{"type":"object","description":"Endereço (JSON)"},"receberNotificacoes":{"type":"boolean","example":true,"description":"Receber notificações"}}},"ChangePasswordDto":{"type":"object","properties":{"currentPassword":{"type":"string","example":"senhaAtual123","minLength":6,"description":"Senha atual"},"newPassword":{"type":"string","example":"novaSenha456","minLength":6,"description":"Nova senha"}},"required":["currentPassword","newPassword"]}}}}